ボタン管理のための基底クラスButtonHavingの追加と、スクロールバーのクラス化の実装とある程度の確認完了
ボタンを押してエディタに起きるアクションは全て基底クラスにButtonHavingを持つように設計する。ButtonHaving内のクラスButtonを継承して新たな独自のボタンクラスを作り、GGEditorのm_buttonsにそのクラスを格納すれば、ボタンが押されたかの判定・押された時の処理・ボタンの描画をGGEditor内に記述しなくても全て行ってくれるようになった。
これを用いて、スクロールバーのクラスを新たに作り(ScrollBar)、それをシステムに取り入れた。四角形内部にマウス点が存在するかは、Crossing Numbers Algorithmを用いて実装した。ただし、エディタ使用の上では不便が生じないため、辺上の判定は曖昧なままにしている。
また、一部のwarning箇所を、warningが発生しないように型キャストするように変更した。

+++ SSR_Collaboration_BU/SSR_Collaboration_BU/BattleObject.cpp
@@ -10,7 +10,7 @@ void BattleObject::ShapeDraw(unsigned int color,int fillFlag,int dx,int dy)const
 }
 
 void BattleObject::PosDraw(unsigned int color,int fillFlag,int size,int dx,int dy)const{
-	DrawCircle(dx+m_pos.x,dy+m_pos.y,size,color,fillFlag);
+	DrawCircle(dx+(int)m_pos.x,dy+(int)m_pos.y,size,color,fillFlag);
 }
 
 void BattleObject::Move(Vector2D v){
